Add AuthenticationRefresher impl
